+++ skia-safe/src/effects.rs
@@ -4,45 +4,31 @@ pub use _1d_path_effect::*;
 mod _2d_path_effect;
 pub use _2d_path_effect::*;
 
-pub mod alpha_threshold_filter;
-pub mod arithmetic_image_filter;
-pub mod blur_image_filter;
-pub mod color_filter_image_filter;
 pub mod color_matrix;
 pub use color_matrix::ColorMatrix;
+
 pub mod color_matrix_filter;
-pub mod compose_image_filter;
 pub mod corner_path_effect;
 pub mod dash_path_effect;
 pub mod discrete_path_effect;
-pub mod displacement_map_effect;
 pub mod gradient_shader;
+
 pub mod high_contrast_filter;
 pub use high_contrast_filter::{high_contrast_config, HighContrastConfig};
+
 pub mod image_filters;
-pub mod lighting_image_filter;
 pub mod luma_color_filter;
-pub mod magnifier_image_filter;
-pub mod matrix_convolution_image_filter;
-pub mod merge_image_filter;
-
-mod morphology_image_filter;
-pub use morphology_image_filter::*;
-
-pub mod offset_image_filter;
 
 mod op_path_effect;
 pub use op_path_effect::*;
 
 pub mod overdraw_color_filter;
-pub mod paint_image_filter;
 pub mod perlin_noise_shader;
-pub mod picture_image_filter;
+
 pub mod runtime_effect;
 pub use runtime_effect::RuntimeEffect;
+
 pub mod shader_mask_filter;
 pub mod stroke_and_fill_path_effect;
 pub mod table_color_filter;
-pub mod table_mask_filter;
-pub mod tile_image_filter;
 pub mod trim_path_effect;
