--- skia-safe/src/effects/compose_image_filter.rs
@@ -1,9 +1,0 @@
-use crate::prelude::*;
-use crate::{image_filters, ImageFilter};
-use skia_bindings::SkImageFilter;
-
-impl RCHandle<SkImageFilter> {
-    pub fn compose(outer: impl Into<ImageFilter>, inner: impl Into<ImageFilter>) -> Option<Self> {
-        image_filters::compose(outer, inner)
-    }
-}
