+++ skia-safe/src/effects/image_filters.rs
@@ -574,6 +574,338 @@ pub fn spot_lit_specular(
     })
 }
 
+impl ImageFilter {
+    pub fn alpha_threshold<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        region: &Region,
+        inner_min: scalar,
+        outer_max: scalar,
+    ) -> Option<Self> {
+        alpha_threshold(
+            region,
+            inner_min,
+            outer_max,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn arithmetic<'a>(
+        inputs: impl Into<ArithmeticFPInputs>,
+        background: impl Into<Option<Self>>,
+        foreground: impl Into<Option<Self>>,
+        crop_rect: impl Into<Option<&'a IRect>>,
+    ) -> Option<Self> {
+        let inputs = inputs.into();
+        arithmetic(
+            inputs.k[0],
+            inputs.k[1],
+            inputs.k[2],
+            inputs.k[3],
+            inputs.enforce_pm_color,
+            background,
+            foreground,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn blur<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        sigma: (scalar, scalar),
+        tile_mode: impl Into<Option<crate::TileMode>>,
+    ) -> Option<Self> {
+        blur(sigma, tile_mode, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn color_filter<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        cf: impl Into<ColorFilter>,
+    ) -> Option<Self> {
+        color_filter(cf, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn compose(outer: impl Into<ImageFilter>, inner: impl Into<ImageFilter>) -> Option<Self> {
+        compose(outer, inner)
+    }
+
+    pub fn displacement_map_effect<'a>(
+        channel_selectors: (ColorChannel, ColorChannel),
+        scale: scalar,
+        displacement: impl Into<ImageFilter>,
+        color: impl Into<ImageFilter>,
+        crop_rect: impl Into<Option<&'a IRect>>,
+    ) -> Option<Self> {
+        displacement_map(
+            channel_selectors,
+            scale,
+            displacement.into(),
+            color,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn distant_lit_diffuse_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        direction: impl Into<Point3>,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        kd: scalar,
+    ) -> Option<Self> {
+        distant_lit_diffuse(
+            direction,
+            light_color,
+            surface_scale,
+            kd,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn point_lit_diffuse_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        location: impl Into<Point3>,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        kd: scalar,
+    ) -> Option<Self> {
+        point_lit_diffuse(
+            location,
+            light_color,
+            surface_scale,
+            kd,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn spot_lit_diffuse_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        location: impl Into<Point3>,
+        target: impl Into<Point3>,
+        specular_exponent: scalar,
+        cutoff_angle: scalar,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        kd: scalar,
+    ) -> Option<Self> {
+        spot_lit_diffuse(
+            location,
+            target,
+            specular_exponent,
+            cutoff_angle,
+            light_color,
+            surface_scale,
+            kd,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn distant_lit_specular_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        direction: impl Into<Point3>,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        ks: scalar,
+        shininess: scalar,
+    ) -> Option<Self> {
+        distant_lit_specular(
+            direction,
+            light_color,
+            surface_scale,
+            ks,
+            shininess,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn point_lit_specular_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        location: impl Into<Point3>,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        ks: scalar,
+        shininess: scalar,
+    ) -> Option<Self> {
+        point_lit_specular(
+            location,
+            light_color,
+            surface_scale,
+            ks,
+            shininess,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn spot_lit_specular_lighting<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        location: impl Into<Point3>,
+        target: impl Into<Point3>,
+        specular_exponent: scalar,
+        cutoff_angle: scalar,
+        light_color: impl Into<Color>,
+        surface_scale: scalar,
+        ks: scalar,
+        shininess: scalar,
+    ) -> Option<Self> {
+        spot_lit_specular(
+            location,
+            target,
+            specular_exponent,
+            cutoff_angle,
+            light_color,
+            surface_scale,
+            ks,
+            shininess,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn magnifier<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        src_rect: impl AsRef<Rect>,
+        inset: scalar,
+    ) -> Option<Self> {
+        magnifier(src_rect, inset, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    #[allow(clippy::too_many_arguments)]
+    pub fn matrix_convolution<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        kernel_size: impl Into<ISize>,
+        kernel: &[scalar],
+        gain: scalar,
+        bias: scalar,
+        kernel_offset: impl Into<IPoint>,
+        tile_mode: crate::TileMode,
+        convolve_alpha: bool,
+    ) -> Option<Self> {
+        matrix_convolution(
+            kernel_size,
+            kernel,
+            gain,
+            bias,
+            kernel_offset,
+            tile_mode,
+            convolve_alpha,
+            self,
+            crop_rect.into().map(|r| r.into()),
+        )
+    }
+
+    pub fn merge<'a>(
+        filters: impl IntoIterator<Item = Option<Self>>,
+        crop_rect: impl Into<Option<&'a IRect>>,
+    ) -> Option<Self> {
+        merge(filters, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn dilate<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        radii: (scalar, scalar),
+    ) -> Option<Self> {
+        dilate(radii, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn erode<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        radii: (scalar, scalar),
+    ) -> Option<Self> {
+        erode(radii, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn offset<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+        delta: impl Into<Vector>,
+    ) -> Option<Self> {
+        offset(delta, self, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn from_paint<'a>(paint: &Paint, crop_rect: impl Into<Option<&'a IRect>>) -> Option<Self> {
+        self::paint(paint, crop_rect.into().map(|r| r.into()))
+    }
+
+    pub fn from_picture<'a>(
+        picture: impl Into<Picture>,
+        crop_rect: impl Into<Option<&'a Rect>>,
+    ) -> Option<Self> {
+        self::picture(picture, crop_rect)
+    }
+
+    pub fn tile(self, src: impl AsRef<Rect>, dst: impl AsRef<Rect>) -> Option<Self> {
+        tile(src, dst, self)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct ArithmeticFPInputs {
+    pub k: [f32; 4],
+    pub enforce_pm_color: bool,
+}
+
+impl From<([f32; 4], bool)> for ArithmeticFPInputs {
+    fn from((k, enforce_pm_color): ([f32; 4], bool)) -> Self {
+        ArithmeticFPInputs {
+            k,
+            enforce_pm_color,
+        }
+    }
+}
+
+impl ArithmeticFPInputs {
+    pub fn new(k0: f32, k1: f32, k2: f32, k3: f32, enforce_pm_color: bool) -> Self {
+        Self {
+            k: [k0, k1, k2, k3],
+            enforce_pm_color,
+        }
+    }
+}
+
+impl Paint {
+    pub fn as_image_filter<'a>(
+        &self,
+        crop_rect: impl Into<Option<&'a IRect>>,
+    ) -> Option<ImageFilter> {
+        paint(self, crop_rect.into().map(|r| r.into()))
+    }
+}
+
+impl Picture {
+    pub fn as_image_filter<'a>(
+        &self,
+        crop_rect: impl Into<Option<&'a Rect>>,
+    ) -> Option<ImageFilter> {
+        self.clone().into_image_filter(crop_rect)
+    }
+
+    pub fn into_image_filter<'a>(
+        self,
+        crop_rect: impl Into<Option<&'a Rect>>,
+    ) -> Option<ImageFilter> {
+        picture(self, crop_rect)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::{CropRect, Dither};
