--- skia-safe/src/effects/tile_image_filter.rs
@@ -1,9 +1,0 @@
-use crate::prelude::*;
-use crate::{image_filters, Rect};
-use skia_bindings::SkImageFilter;
-
-impl RCHandle<SkImageFilter> {
-    pub fn tile(self, src: impl AsRef<Rect>, dst: impl AsRef<Rect>) -> Option<Self> {
-        image_filters::tile(src, dst, self)
-    }
-}
